Make the message trimming behaviour consistent across platform (#936)
* refactor: remove old trim behaviour and create new defaults

* refactor: add in trimming by before time

* refactor: remove old trim thread job, remove message processor scheduling trim thread

* refactor: remove spacing

* refactor: enable trimming by default

* refactor: remove trim now option in chat preference

+++ app/src/main/java/org/thoughtcrime/securesms/database/MmsSmsDatabase.java
@@ -162,8 +162,8 @@ public class MmsSmsDatabase extends Database {
     }
   }
 
-  public int getConversationCount(long threadId) {
-    int count = DatabaseComponent.get(context).smsDatabase().getMessageCountForThread(threadId);
+  public long getConversationCount(long threadId) {
+    long count = DatabaseComponent.get(context).smsDatabase().getMessageCountForThread(threadId);
     count    += DatabaseComponent.get(context).mmsDatabase().getMessageCountForThread(threadId);
 
     return count;
